OriginalDocSub, ATC Fixes

+++ FintrakBanking.Entities/Models/TBL_ATC_LODGMENT.cs
@@ -40,7 +40,7 @@ namespace FintrakBanking.Entities.Models
 
         public int STATUSID { get; set; }
 
-        public int APPROVALSTATUSID { get; set; }
+        public short APPROVALSTATUSID { get; set; }
 
         public int CREATEDBY { get; set; }
         public DateTime DATETIMECREATED { get; set; }
